CLN: Reorganize cursor
