+++ pymapd/_parsers.py
@@ -1,6 +1,58 @@
 """
 Utility methods for parsing data returned from MapD
 """
+from collections import namedtuple
+import mapd.ttypes as T
+
+
+Description = namedtuple("Description", ["name", "type_code", "display_size",
+                                         "internal_size", "precision", "scale",
+                                         "null_ok"])
+
+_typeattr = {
+    'SMALLINT': 'int',
+    'INT': 'int',
+    'BIGINT': 'int',
+    'TIME': 'int',
+    'TIMESTAMP': 'int',
+    'DATE': 'int',
+    'BOOL': 'int',
+    'FLOAT': 'real',
+    'DECIMAL': 'real',
+    'DOUBLE': 'real',
+    'STR': 'str',
+}
+
+
+def _extract_row_val(desc, val):
+    # type: (T.TColumnType, T.TDatum) -> Any
+    typename = T.TDatumType._VALUES_TO_NAMES[desc.col_type.type]
+    return getattr(val.val, _typeattr[typename] + '_val')
+
+
+def _extract_col_vals(desc, val):
+    # type: (T.TColumnType, T.TColumn) -> Any
+    typename = T.TDatumType._VALUES_TO_NAMES[desc.col_type.type]
+    return getattr(val.data, _typeattr[typename] + '_col')
+
+
+def _extract_description(row_desc):
+    # type: (List[T.TColumnType]) -> List[Description]
+    """
+    Return a tuple of (name, type_code, display_size, internal_size,
+                       precision, scale, null_ok)
+
+    https://www.python.org/dev/peps/pep-0249/#description
+    """
+    return [Description(col.col_name, col.col_type.type,
+                        None, None, None, None,
+                        col.col_type.nullable)
+            for col in row_desc]
+
+
+def _is_columnar(data):
+    # type: (T.TQueryResult) -> bool
+    return data.row_set.is_columnar
 
 
 def _load_schema(buf):
